Added testing commands

+++ src/main/java/me/marquez/bettercinematics/utils/FileUtils.java
@@ -13,25 +13,34 @@ public class FileUtils {
 
     private static final Gson gson = new Gson();
 
-    public static <T> CompletableFuture<T> loadFromJson(@NonNull File file, @NonNull Class<? extends T> clazz) {
+    public static <T> T loadFromJson(@NonNull File file, @NonNull Class<? extends T> clazz) throws IOException {
+        return gson.fromJson(JsonParser.parseReader(new FileReader(file)), clazz);
+    }
+
+    public static <T> CompletableFuture<T> loadFromJsonAsync(@NonNull File file, @NonNull Class<? extends T> clazz) {
         CompletableFuture<T> future = new CompletableFuture<>();
         Executors.newCachedThreadPool().submit(() -> {
             try {
                 future.complete(gson.fromJson(JsonParser.parseReader(new FileReader(file)), clazz));
-            } catch (FileNotFoundException e) {
+            } catch (Exception e) {
                 future.completeExceptionally(e);
             }
         });
         return future;
     }
 
-    public static <T> CompletableFuture<Void> saveToJson(@NonNull File file, @NonNull Class<? extends T> clazz, @NonNull Object object) {
+    public static <T> CompletableFuture<Void> saveToJsonAsync(@NonNull File file, @NonNull Class<? extends T> clazz, @NonNull Object object) {
         CompletableFuture<Void> future = new CompletableFuture<>();
         Executors.newCachedThreadPool().submit(() -> {
             try {
+                if(!file.exists()) {
+                    File parentFile = file.getParentFile();
+                    if(!parentFile.exists() && !parentFile.mkdirs()) throw new IOException("Can not make directories");
+                    if(!file.createNewFile()) throw new IOException("Can not create new file");
+                }
                 gson.toJson(object, clazz, new FileWriter(file, true));
                 future.complete(null);
-            } catch (IOException e) {
+            } catch (Exception e) {
                 future.completeExceptionally(e);
             }
         });
